Changed the datatype of alpha and beta parameters to float

+++ vx_nn/src/convolution_layer.cpp
@@ -24,8 +24,8 @@ THE SOFTWARE.
 
 struct ConvolutionLayerLocalData {
     NeuralNetworkCommonHandle * handle;
-    double alpha;
-    double beta;
+    float alpha;
+    float beta;
     miopenTensorDescriptor_t input_desc;
     cl_mem input_mem;
     miopenTensorDescriptor_t weight_desc;
