+++ vx_nn/src/deconvolution_layer.cpp
@@ -31,8 +31,8 @@ THE SOFTWARE.
 
 struct DeconvolutionLayerLocalData {
     NeuralNetworkCommonHandle * handle;
-    double alpha;
-    double beta;
+    float alpha;
+    float beta;
     miopenTensorDescriptor_t input_desc;
     cl_mem input_mem;
     miopenTensorDescriptor_t weight_desc;
